# backend/swagger.json
{
  "openapi": "3.0.0",
  "components": {
    "schemas": {
      "Id": {
        "type": "string",
        "minLength": 24,
        "maxLength": 24,
        "pattern": "^\\d{24}$"
      },
      "MangaUpdateReport": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "site": {
            "type": "string"
          },
          "link": {
            "type": "string",
            "format": "url"
          },
          "newChapCount": {
            "type": "integer"
          }
        }
      },
      "Chapter": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "name": {
            "type": "string"
          },
          "link": {
            "type": "string",
            "format": "uri"
          },
          "isRead": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Manga": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "user": {
            "$ref": "#/components/schemas/Id"
          },
          "name": {
            "type": "string"
          },
          "link": {
            "type": "string",
            "format": "uri"
          },
          "site": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "chapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Chapter"
            }
          },
          "newChapCount": {
            "type": "integer"
          },
          "unreadChapCount": {
            "type": "integer"
          },
          "isCompleted": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "minimum": 0,
            "maximum": 3
          },
          "statusString": {
            "type": "string",
            "enum": [
              "finished",
              "last chap reached",
              "many to read",
              "new chap"
            ]
          },
          "shelf": {
            "type": "string",
            "enum": [
              "to read",
              "reading",
              "waiting",
              "dropped",
              "finished"
            ]
          },
          "note": {
            "type": "string"
          },
          "hidden": {
            "type": "boolean"
          },
          "lastReleased": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "info": {
    "title": "MangaBookmark API",
    "version": "3.0.0",
    "description": "API for MangaBookmark client"
  },
  "paths": {
    "/api/mangas": {
      "get": {
        "description": "Search & filter mangas",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "description": "Search mangas by name",
            "schema": {
              "type": "string",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "shelf",
            "schema": {
              "type": "array",
              "items": {
                "enum": [
                  "to read",
                  "reading",
                  "waiting",
                  "dropped",
                  "finished"
                ]
              },
              "required": false
            }
          },
          {
            "in": "query",
            "name": "isCompleted",
            "schema": {
              "type": "boolean",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "minimum": 0,
                "maximum": 3
              },
              "required": false
            }
          },
          {
            "in": "query",
            "name": "hidden",
            "schema": {
              "type": "boolean",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "site",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "required": false
            }
          },
          {
            "in": "query",
            "name": "createdAtGTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "createdAtLTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "lastReleasedGTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "lastReleasedLTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "unreadChapCountGTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "unreadChapCountLTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "required": false
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "If set to 0, all pages are retrieved.",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "The field name to sort. Add - to sort desc. This field can be a string of one filed a string of multiple fields separated by space, or an array of string.",
              "required": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Manga"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "totalItem": {
                      "type": "integer"
                    },
                    "totalPage": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Create new manga",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "link": {
                    "type": "string",
                    "format": "uri"
                  },
                  "readChapters": {
                    "type": "array",
                    "required": false,
                    "items": {
                      "type": "string",
                      "format": "url"
                    }
                  },
                  "note": {
                    "type": "string",
                    "required": false
                  },
                  "isCompleted": {
                    "type": "boolean",
                    "required": false
                  },
                  "hidden": {
                    "type": "boolean",
                    "required": false
                  },
                  "shelf": {
                    "type": "string",
                    "enum": [
                      "to read",
                      "reading",
                      "waiting",
                      "dropped",
                      "finished"
                    ],
                    "required": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Manga created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Manga"
                }
              }
            }
          }
        }
      }
    },
    "/api/mangas/{mangaId}": {
      "patch": {
        "description": "Edit manga",
        "parameters": [
          {
            "in": "path",
            "name": "mangaId",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "note": {
                    "type": "string",
                    "required": false
                  },
                  "isCompleted": {
                    "type": "boolean",
                    "required": false
                  },
                  "hidden": {
                    "type": "boolean",
                    "required": false
                  },
                  "shelf": {
                    "type": "string",
                    "enum": [
                      "to read",
                      "reading",
                      "waiting",
                      "dropped",
                      "finished"
                    ],
                    "required": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Manga edited successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Manga"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete manga",
        "parameters": [
          {
            "in": "path",
            "name": "mangaId",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          }
        }
      }
    },
    "/api/mangas/info": {
      "get": {
        "description": "Get live manga info. This manga might NOT be in the database.",
        "parameters": [
          {
            "in": "query",
            "name": "link",
            "schema": {
              "type": "string",
              "format": "uri",
              "required": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get info successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "link": {
                      "type": "string",
                      "format": "uri"
                    },
                    "image": {
                      "type": "string",
                      "format": "uri"
                    },
                    "isCompleted": {
                      "type": "boolean"
                    },
                    "chapters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "link": {
                            "type": "string",
                            "format": "uri"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/mangas/{mangaId}/mark-chapters": {
      "patch": {
        "description": "Update chapter isRead status",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isRead": {
                    "type": "boolean",
                    "required": true
                  },
                  "chapters": {
                    "type": "array",
                    "required": true,
                    "items": {
                      "type": "string",
                      "format": "uri"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chapters marked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Manga"
                }
              }
            }
          }
        }
      }
    },
    "/api/mangas/{mangaId}/update": {
      "post": {
        "description": "Update this manga",
        "parameters": [
          {
            "in": "path",
            "name": "mangaId",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Manga"
                }
              }
            }
          }
        }
      }
    },
    "/api/mangas/update-multiple": {
      "get": {
        "description": "Update mangas found by search & filters",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "description": "Search mangas by name",
            "schema": {
              "type": "string",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "shelf",
            "schema": {
              "type": "array",
              "items": {
                "enum": [
                  "to read",
                  "reading",
                  "waiting",
                  "dropped",
                  "finished"
                ]
              },
              "required": false
            }
          },
          {
            "in": "query",
            "name": "isCompleted",
            "schema": {
              "type": "boolean",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "minimum": 0,
                "maximum": 3
              },
              "required": false
            }
          },
          {
            "in": "query",
            "name": "hidden",
            "schema": {
              "type": "boolean",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "site",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "required": false
            }
          },
          {
            "in": "query",
            "name": "createdAtGTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "createdAtLTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "lastReleasedGTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "lastReleasedLTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "unreadChapCountGTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "unreadChapCountLTE",
            "schema": {
              "type": "string",
              "format": "date",
              "required": false
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "required": false
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "If set to 0, all pages are retrieved.",
              "required": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "description": "number of updated mangas"
                    },
                    "success": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MangaUpdateReport"
                      }
                    },
                    "fail": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MangaUpdateReport"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}